requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
celery>=5.3.6
redis>=5.0.1
requests>=2.31.0
django-cors-headers>=4.3.1 